pyproject.toml
[project]
name = "livesub"
version = "0.1.0"
description = "Real-time transcription overlay for Linux desktop audio"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.4.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "PyQt6>=6.6.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11.0",
]

[project.scripts]
livesub = "live_transcript:main"
